Took a bit longer than expected. This adds the requested amazon AR -> albummeta feature. Newly created, modified or deleted Amazon ARs will directly affect the asin and coverarturl collumns in album_amazon_asin and albummeta. The coverarturl was kept in the DB, because it helps in keeping a lot of scripts clean from special AR handling.
Most of the new functionality is provided by Album::ParseAmazonURL($url), which takes any url and extracts the asin and coverarturl into the album object, and Album::UpdateAmazonData($mode), which allows inserts, removes and updates of amazon data.

Further cleaned up the mason scripts, that parsed the urls themselves; removed the ALBUM_AMAZON_ASIN_LINKID constant, because it was wrong and should be read from the database (get it with Album::GetAsinLinkTypeId now); added an upgrade notes to UPGRADE.

Added an update directory to admin, that should be used to store one-time upgrade scripts like the two I added for the failed moderation count fix and the ASIN AR -> albummeta change.