IMPORTANT:
I keep accidentally checking in DBDefs.pm with local information since we need to modify
this for a server configuration. Each check in I have one extra step to do to make
sure that I don't commit those changes. 

From now on the default settings of DBDefs.pm will be in DBDefs.pm.default and the person
installing a copy of mb_server will need to copy DBDefs.pm.default to DBDefs.pm before
configuring DBDefs.pm. 

This should prevent accidental check-ins of DBDefs.pm and it's more consistent with
the principles of version control. IMHO.