Replace Utils::deep_equal with a recursive solution
Storable does not seem to be working anymore, as the strings always end up a
few characters out, which I think could be an encoding issue. For now, I have
replaced it with a custom recursive scan search. We could memoize it if it
becomes a bottleneck of sorts.

This fixes MBS-118