Started work on rewriting CoverArt.pm
This has been split into a few different entities. MB::S::CoverArt
is like an entity, and is attached to releases when they are loaded
with the CoverArt data object. MB::S::CoverArt::Provider::* is a set
of cover art providers (currently just Amazon and a regular expression
based one), which will allow us to trivially hook in new providers.

This commit completes the regular expression parser, and tests it.