Prevent creating a relationship between 2 identical entities.
This is a very basic check that makes sure the source and destination
entities have different IDs. But it doesn't prevent you from actually
getting here.

We really need to think about the proper way to create relationships,
the way I've done it was just hacked together because I didn't have time
to create JavaScript query boxes and such.