Add quality attribute for artists & releases
Adds a quality column to the schema for artists and releases, which is
what we have on the current servers. Entities and data accessors have
been updated to bring in this column.

Upgrade NGS upgrade scripts to copy over the quality column from the
older schema as-is.

Unit tested entities and data accessors