Add support for the Amazon product advertising API.
This uses the Amazon web services to fetch cover art. The API seems to have changed
from the old AWS system, and now requires a public and private key to sign requests.
This data is stored in DBDefs, so will require a DBDefs update to test this.

The cover art resolution is fairly naive at the moment - it only uses the medium
sized image from the primary cover art source, ignoring everything else.