Rewrote Countdown plugin to keep translation in the templates
This method might be hackish, but I think it certainly does the job very well and keeps everything flexible, and more importantly, on the template side of things.
Further improvements (if necessary) would be to include count*up* support (i.e. "5 hours ago") and also move the expiration_time BLOCK to common-macros.tt.