Fix an edge case when merging relationships.
If there are 3 entities being merged (A and B are being merged into C)
then relationship merging would fail if A and B shared a common
relationship that was *not* present in C.

The previous relationship merging logic only filtered relationships that
were common between A and C, and B and C, which would lead to a unique
constraint violation if A and B shared the same relationship.

This logic has been changed now to a 2 step process of first filtering
for unique relationships, ignoring the source entity, and then changing
the source entity for *all* these relationships to the merge target.