List all artists related to a work on /artist/works.
Artists shown are those artists with a direct AR to the work in question, or
artists that have a relationship to a recording that is related to the
work. Artists with direct ARs are shown first.

Artists that are indirect (via recordings) are filtered to only show artists in
recording artist credits for the 3 most popular recordings, where popularity is
measured by the amount of ARs the recording has.