Test all Data objects for memory cycles.
As Perl uses reference counting to handle garbage collection, it's easy
to end up creating cyclic data structures. For the data objects
themselves, this isn't really a problem (they are effectively
singletons). However, if we introduce cycles in MB::S::Entity objects
they we leak them through the life of the web server.

This commit adds memory cycle tests for just about everything where we
had existing tests. I've tried to fix them in a sane way, but sadly some
of the code we have is pretty much designed to be cyclic so I've had
to resort to weakening by breaking encapsulation.