If the tracklist has changed after editing, reject the edit.
Assume an editor A edits a medium and changes the tracklist, and creates a
pending edit. While this edit is open, editor B edits the same medium, changing
the tracklist as an auto-edit. As editor A's edit has now been superseded, the
edit is rejected by ModBot.

Note that this logic is only concerned with tracklist information for now and
NOT other data.