Rewrite EmptyArtists.PL to use a looping algorithm
After running for over an hour, EmptyArtists.PL had not completed,
and the cost was so high, I think we can do better.

This replaces it with an iterative implementation, that loops over
each artist (in this case, each artist that was updated within the
last $interval). If that artist has no relationships, or artist credits
attached, it is then returned for removal.

The rest of the EmptyArtists script is the same.