Decode HTML entities when loading annotations
We should prefer not to store HTML in the database at all, but some of the old
data does have HTML entities in. This will get stripped when we add new
annotations, but we don't want to create annotation edits if the text is
logically the same. Decoding HTML entities when loading annotations makes text
comparisons work as expected.