Add special names for displaying in the sidebar
Displaying the pretty_name in the sidebar might work, but we don't want to start
adding more pretty_name's and screwing up the relationship list. For some sites
like Discogs and Twitter, the name only makes sense if you have the icon or more
context, which is not available in the relationship list.

This introduces the 'URL::Sidebar' role, which requires specific URL's implement
the 'sidebar_name' method.