Make sure all artists in new artist credits exist before applying edits
If an edit is created that uses an artist credit, there is a chance that the
artist will not exist when the edit is applied (for example, it is cleaned up as
any empty artist). We now check that all artists exist before applying the
edit. If any artists do not exist, the edit fails as 'failed dependency'