Make sure all edits have an ID before accepting
Some edits need the id when accepting (for example, to delete subscriptions) but
this is only available if the edit has been inserted. Auto-edits, however, are
applied before insertion so this is not available.

Now we use nextval explicitly to generate the edit ID, and do not rely on the
default behaviour on insertion.