Add AR migration logic for medley and cover track-track ARs
Cover ARs are now migrated depending on the attributes present. If a
track-track cover AR has the 'translated' or 'parody' attributes, the
AR is migrated as a work-work 'other version' relationship.  If these
attributes are not present, the relationship is migrated as a
recording-work 'performance' relationship with the (new) 'cover'
attribute.

Medley's are now migrated as recording-work, but the new work-work
relationship link type still exists, allowing for work-work
relationships in the future.