Split JavaScript & CSS into groups with consistent minification & hashing
JavaScript now lives in 4 groups:

    - common - included on all pages
    - edit - included on pages that do editing
    - guess-case - included on pages that require guess-case
    - release-editor - included in release editor pages

CSS lives in a single group, which is just all our (internal MB) CSS.

Each group has a 'manifest' file (root/static/{name}.js.manifest) which
specifies which files make up the group. A directory name in the
manifest means to take all files recursively, while a filename adds just
that file to the pack.

If the server is running under debug mode (CATALYST_DEBUG { 1 } in DBDefs) all
documents will be served as normal. However, if debug mode is not on, it will
attempt to use the packed resource. This must be generated by running the new
`script/compile_resources.pl` script.