Make sure that link type UUIDs are stable between migrations
When we create a new link type in the migration process, we need to generate a
new UUID for it. Previously, this UUID was created from a URL that used the new
link type id. However, as this ID is automatically generated in an order that is
not guaranteed (because link types aren't migrated in any specific order), the
UUID could change between migrations.

This now uses the link type's name to generate a UUID from, not the ID. The
rewritten link phrases have also had the UUID's updated.