First pass at medium attributes for disc level relationships
Some relationships point to the exact disc, not just the release as a concept -
for example the 'DJ-mixed' relationship. This changes the merge strategy to
first alter medium-level relationships to have a disc attribute (if the release
appears to have multiple discs). These relationships are then later combined
into single relationships, depending on begin/end dates and artists, for release
entity.