Further filter the edit_ids in voting to only votable edits
In enter_votes, the @votes array is filtered to contain only votes for edits
that exist, are open, and that do not currently already contain a non-superseded
vote that is the same as the new vote. However, the @edit_ids array was being
calculated once, and *not* updated. This could lead to some bugs and
unintentional behaviour.

I'm not sure if this yet addresses the multiple voting problem on NGS, but it's
certainly not correct how it was before...