Resolve artist credits to IDs inside tracklist search/creation
If we resolve artist credits in the Edit:::Medium::Edit edit type, we
may find artist credits that will subsequently get removed when
replacing a tracklist. For example, if an artist credit is only used on
a track in a tracklist we are about to replace, when we deleted those
tracks we will decrement the ref count of artist credits to 0, and thus
delete the artist credit.

By moving the artist credits inside the Tracklist data object, we avoid
this scenario.