Merge duplicate tracklists
For some reason or another, we have duplicate tracklists in the system, which is
not by intention. This adds an upgrade script that is capable of merging
duplicate tracklists.

We create a row for every tracklist, which contains an array of track names,
artist credits, recording ids and lengths, all ordered by position. This set of
arrays defines a unique tracklist. Then we find duplicate occurances of these
tracklists, and merge them and remove the duplicate tracklists. The destination
tracklist is selected deterministically as retaining the tracklist with the
lowest ID.