Minimize the tracklist differences display
The tracklist differences display now only shows changes in text displayed. This
means track name changes, artist credit name changes, and track duration
changes. "Invisible" changes, such as a change in a recording link or artist in
an artist credit are no longer part of this diff, and will be displayed
separately.