Fix historic "edit label" edits
They still had an 'entity_id' section, which should actually be a dictionary
called entity containing the entity ID and name. I've attempted to fill this in
by loading labels by ID, but in the case we can't load this data I've used
'[removed]' (which would be displayed in this case anyway).