Avoid auto-vivifying the label section of add release label edits
The release editor always provides $opts{label} but it *might* be
undef. If it's undef, then checking for exists still returns true, which in turn
autovivifies the undefined value into an empty hash reference (when determining
related entities). The check has been changed to check for truth, to avoid this.