Added the 'ID' predicate for edit searching
Starting to sketch out the edit search system. We have a collection of predicate
classes, which represent certain types of search conditions (date searching,
entity searching, enumeration searches, etc). Predicates are collected into a
Query object, which combines all these predicates in a certain way (ie, AND or
OR, though this work will be done in a later commit).