Remove all mention of raw_dbh
raw_dbh is now the same as dbh, so we no longer need to keep track of this
separately. Likewise, raw_sql is now sql, so that can also go. raw_conn
is the same as conn, so that has also been removed. RAWDATA no longer
exists, so references to this have been removed.