Add a simple logging system in
This adds a basic logging system based on Log::Dispatch and inspired by
Log::Contextual. If we need to add logging to something, we just need to
do the following.

    use MusicBrainz::Server::Log qw( log_info );
    # Later...
    log_info { 'Important message here' };

The logging system uses Data::Dumper::Concise, which means formatting messages is a
breeze too - which should mean less barriers to providing *useful* debug
output.

    log_error { "Artist has invalid data: $_" } $artist

The logger is configured through DBDefs, which should allow enough flexibility
to use this logger usefully in a production scenario (for example, logging
through syslogd).