When selecting an artist credit from the main artist input, clear the sub-fields
If the user makes a selection from the main artist box on the release page, and
makes a selection, they will expect the credited name to be the same as the
artist they selected. Essentially, they expect the artist credit to match the
artist they selected.

This now clears the artist credit on selection, which allows it all to be reset
to the selected artist.