Do not show duplicate PUIDs/ISRCs in /ws/1
If a release has the same recording multiple times, then PUIDs and ISRCs would
be attached to the same recording more than once. This happens because the
recordings are exactly the same object (same refaddr). This adds a uniq filter
on refaddr, so we only attach PUIDs and ISRCs to a recording once.