Display standalone recordings in search results
The search results assumed that a recording appeared on 1 or more releases, and
results were displayed by enumerating these releases. Standalone recordings
however have no releases, so this branch of code was never reached (and thus
results were not displayed).

If a recording appears on no releases, we now display the recording distinctly
as a standalone recording in the search results.