Keep track of entities being merged in page data
There is a bizzare sort of race condition with the current merging code. If a
user selects 2 entities to merge, they will be redirected to the merge
confirmation page. However, if they then add another entity for merging in a
separate tab, then this *won't* be reflected on the previous merge confirmation
page, but *will* be used when that page is submitted.

This leads to a problem of trying to merge releases, where the list of releases
being merged does not match where mediums should be placed.

The pages have now been changed to have a list of hidden entity ID fields,
meaning only what the user sees on the page will be merged.