Editing a cover art relationship should update cover art
If a relationship is created, it should immediately rebuild cover art. If this
edit is later rejected, cover art should be rebuild again. When a cover art
relationship is edited, that should also rebuild cover art for this
release. Finally, if a cover art relationship is deleted, we should rebuild
cover art again.