Move the Lucene "OR *" work around to happen around the search
Previously, for enitites that could have aliases, we would be searching for
"name OR name*" over all search fields. However, what we were really *meant* to
be expressing was a search for "name" (using aliases etc) OR a search for
"name*".

This changes the autocompletion role to form the search query twice with "name"
and "name*" and then combine that with OR.