When garbage collecting a tracklist, garbage collect recordings
When a release is removed, it might cause the tracklist to be deleted.
When a tracklist is deleted, all the tracks are removed too, which may
leave some recordings to have no track pointing to them. In some cases,
this is ok (if the recording was created as a standalone recording), but
in some cases it's bad -- for example if the recording was created
through the release editor in the first place.

This adds the new 'garbage_collect_orphans' method to the recording DAO,
which will check for, and delete, recordings which have:

    - No edits pending
    - No tracks
    - No add track, add track kv (both historic) or add standalone
      recording edit