MBS-3301: When editing mediums, do allow some new recordings to be created
The previous check for recordings meant that *all* recording IDs must be
present, though we do allow some to be undefined as this will create a new
recording for them. Now, we count the amount of undefined recordings, and make
sure that number is the same when we enter the edit. If it's not, it means some
of the other recordings have been merged.