Fix test failures found by new Edit::Medium::Edit tests
There was a problem in finding recordings that no longer exist. If a tracklist
contained multiple references to the same recording, the count of fetched
recordings would not match the expected count from the tracklist, and
incorrectly fail. The algorithm has now been changed to find all recordings, and
then check for existing when we build the final tracklist.