Do not allow creating relationships with 'grouping' relationship types
Grouping relationship types do not have a description, and users should not be
able to use this in a relationship. I've split the form up to add a 'LinkType'
role, which checks attributes are within bounds (this checking has been
improved), and also checks that the link type is valid and not a group.