Slightly rewrite the outdated_releases query
This changes the query to select distinct releases, and orders the
distinction to favour relationships first, with ASINs first, and the
most recently updated relationship. If there are no relationships, then
the barcode will be selected. Finally, the distinct set of updates to
perform, is sorted by the least recently updated first.