Use current artist credit ID in checks in Release::EditArtist
When applying the EditArtist edit, we check if any tracks have an artist credit
that matches the release artist. If they are the same, then that track needs to
be updated.

Previously, the edit was doing a find or insert for the old artist credit. This
will fail though, if any of the artists in the old artist credit have been
merged, as the old_artist_credit data section will be invalid. Instead, I just
use the current artist credit, and entirely ignore the old artist credit when
applying the edit.