Do not try to merge recordings when there are multiple targets
When merging an old release into a new one, and merging recordings,
we should only merge recordings if it's clear and unambiguous what the
new recording is. If an old recording corresponds to multiple new
recording IDs, then we don't really know how to merge and we should
continue without merging that.