MBS-3378: Clean up all Integer fields by using 0+ to coerce to integer
http://musicbrainz.org/edit/15080650 appears to show no changes, because it
changes the begin date from 2002-9-2 to 2002-09-02 (only adding 0s). However,
these are the same date, so it really shouldn't have been allowed in the first
place. This happened because we were using string equality checks on different
strings.

This commit adds the new MB Integer field, which simply does 0+ on the input
(after previous validation), to turn the string into an integer.