Rewrite OtherLookup processing to be more general
I've introduced a controller-specific DSL for the other lookup processing. The
'lookup_handler' function adds a type of lookup handler. This first does some
basic form validation, and then hands the result it got to a specific handler
which does some work (usually just dispatching a redirect, but sometimes a
direct lookup).

This simplifies the controller somewhat, and allows for a nice simplification in
the template, where each field is it's own form. This is also clearer, as before
the processing would take the first non-empty field you specified. Also, users
may have been under the impression they could combine fields, which is not true.