Rewrite TagLookup to do a more intelligent search
This rewrites the TagLookup code to do a proper external search. The changes
are:

* Different types of searches depending on what is input to the search. If the
  user give all fields, then we are able to do a track level search. If the user
  only provides an artist name, then we can only search for artists.

* Correctly handle artist aliases.

* Fix a minor bug with the form naming.

This will be better when MBS-2096 is fixed, and we have a clear distinction
between searching for tracks and searching for recordings. A search for track
should give one row per track, and each row should have a 'tagger' button for
releases.