Use Encode::Detect as a fallback if we can't work out what encoding a parameter is in
If you try to go to /?q=%FC for example, the app used to
crash. Catalyst::Plugin::Encoding::Unicode now provides a callback in the case
of an exception, so the app gets a chance to handle it. We now make use of this
callback, and use Encode::Detect as a last attempt. If that still fails, then we
do crash, but I imagine this won't happen as often as before.