Make sure editing or creating a relationship is atomic with the existance check
The check for a relationships existance and the creation of a relationship
should be entirely atomic; that is - it should be impossible to interleave these
operations between 2 processes. The absence of a lock however, meant that it was
possible - and people were hitting this problem by accidently submitting a form
twice.

This commit mostly introduces the lock_and_do method of the relationship data
object, but also does necessary refactoring in the relationships controller in
order to make sure all relationship editing code goes through this call.