Attempt to merge partial dates, when the target has no partial date at all
If the target does not have a partial date, it might be possible to salvage some
of the attributes from the merge sources. When merging, the strategy is to find
the most useful partial date. The general priority is:

* Has a year, month and day
* Has a year and month
* Has just a year

This means if there's a choice of taking just the year from one entity, or just
the month from another, it'll take just the year. I consider the year to be of
more interest than a month.