MBS-1947: Do not 404 on private collections
Instead, request a login if there is no user, and 403 if there is a user but it
is not the collection owner. I'm not entirely happy with this, as it does still
reveal valid collection MBIDs, but it seemed to be agreed on in IRC [1].

I've also introduced the 'own_collection' chained action, which means that
add/remove/edit actions are only valid on your own collection (rather than just
requiring a user is logged in).

[1]: http://chatlogs.musicbrainz.org/musicbrainz-devel/2011/2011-06/2011-06-17.html#T13-35-30-278265