Attempt to fix historic 'move disc ID' edits
These edits have been incorrectly migrated, and are lacking 'previous
release' IDs. This commit attempts to fix that by re-migrating that
portion of these edits. It seems that 'prevvalue' is the correct field
to use here.

It attempts to relink to the resolved release, but this might not always
work. For example, if that old release has since been merged, the
resolver will point to a non-existing release. In this case, we can do
nothing more but shrug and move on.