MBS-3545: Make the tag lookup page able to search again
My "improvements" meant it redirected to an external search, using the special
"alias" query, but this makes it really difficult for users to search
again. Now, it forwards to the external search logic, but displays the results
itself and display the same form as before. This allows a user to change their
search query, but still have the power of an external search to do the
searching.