MBS-3559: Fix works searches
The transformer from the search schema to the internal MusicBrainz schema was
extremely fragile to input, as it used (keys %hash)[0] to refer to a specific
key. However, with the fix of SEARCH-106, a new 'target-type' key was added,
causing this code to now break. I've fixed the code to not rely on the elements
in the hash, and instead make use of this new target-type attribute.