Make sure that release editor 'edits' JSON is escaped in the HTML
ee31633e adds 'fif_from_value => 1' to the edits field, which means that
it's value will be passed through verbatim into the output HTML. This
means that if there is a literal &amp; that a user wants to display in
the tracklist, this gets turned into just '&'. By removing this setting,
&amp; can be used as expected.

Sadly, ee31633e does not mention why this was added, but as there is no
special handling of escaping. This does open up a risk of cross-site
attacks, especially as the release editor is meant to be extendable by
other people writing scripts to submit to it.