MBS-3528: Fix moving relationships between recordings
The 'EntityAutocomplete' code attempts to be general by enumerating all
properties of 'data.item' (the results of an autocomplete search) and setting
the value of input elements with the same class as each properties name, to the
property value. However, it was using $.each, and $.each has special logic if
the passed in object to enumerate has a 'length' property:

   Arrays and *array-like objects with a length property* (emphasis added)

Recording searches do return objects with a length property, but that's the
length of the recording. To work around this, I've replaced the call to $.each
with a manual loop (copied almost verbatim from jQuery).