When loading relationships for /ws/1 entities, load work level artist rels too
This commit loads all work-artist relationships for entities that have relationships,
and then displays these relationships as if they were from the original source
entity. That is, if for a recording->work->artist relationship, a
recording->artist relationship is displayed. While not strictly true, this does
give back-compat to /ws/1 (see MBS-3336 for the specific problem).