Do not show duplicate artists by alias searches, if the artist name alone was a duplicate
When finding duplicate artists, we first search by name, and then do a second
search for aliases. This alias search however would replace a duplicate entry
even if it already existed. It's more useful if we keep the original artist name
duplicates however, so the code has been changed to only add a duplicate artist
if we haven't seen that name/gid combination.