MBS-3656: Fix the autoeditor 'vote-on-all' logic to handle edge cases
The 'vote-on-all' widget is created by cloning the first set of voting options
on the page. As there are pages that might show a mix of open and closed edits,
this is not necessarily the very first row. The vote-on-all widget radio buttons
should not have an in-use ID, but the .replace() call was too specific. This has
been replaced with a more general regular expression.