MBS-3657: Fix can_merge for append
The idea between the can_merge check was to take all medium position
changes, and complement that set with any mediums that haven't been
accounted for. We then need to make sure that these sets don't have any
conflicting positions. The computation of the second complementing set
however was wrong, which meant it would incorrectly believe to have
found conflicts.