Add a Release::AddCoverArt edit type
This edit type assumes that a pending-{timestamp} image has been uploaded to the
correct bucket on the coverartarchive.org. On acceptance, it moves this
pending-{timestamp} image to the correct location, possibly deleting already
existing artwork in that location.

If the edit is rejected, then the pending artwork is deleted.