Correctly namespace the custom types we define
Net::Amazon::S3 is not a good citizen, and pollutes the global type namespace
with a 'DateTime' type. We're also not a good citizen, as we do the same - and
this leads to a conflict in type definitions.

This commit correctly namespaces our custom type definitions using
MooseX::Types, in the hope of preventing this in the future. All constants in
::Types have been moved to ::Constants.