Move the URL cleanup code to database triggers
Rather than having the logic for when to try and remove a URL be in the Perl
code, move it to the database. This happens to now catch the case when an entity
is deleted and all its relationships are subsequently deleted. It also keeps the
Perl code nice and simple.