Fix processing bug with cover art script
The SQL to select releases to update should find relationships first, and if
this fails fall back to the barcode. A slight logic error however meant that if
a release had URL relationships that would not provide cover art, these could
possible take precendance over correct cover art relationships, because the row
would have a barcode.

This happened because the _sort_order field is sorted descending (so the row
with the highest sort order should be selected). PostgreSQL defaults to sorting
NULLs first when sort is order is DESC, so the discogs/barcode relationships
were being sorted before everything else. This has been changed to sort null last.

Erring on the side of caution, I have also modified the join condition to only
join relationships that are within the approved types for l_release_url
relationships that provide cover art.