Remove DISTINCT clause from find_by_recordings
If you have a search for a recording and it matches multiple recordings on the
same release, only the first recording row would have release information;
subsequent recordings would display as standalone (which is wrong).

We could expand the distinction clause to be over both recording.id and
release.id, but removing it entirely means that if a single recording appears
multiple times on a release, the search results will reflect this. I consider
this a useful feature, so I've left it in.