Fix URL editing that causes a merge
If you edit a URL into a URL that already exists, then the 2 are
merged. However, commit 02a2a416782 introduced database triggers that meant when
a URL had no relationships, it was deleted. This state is encountered in the
merge process when relationships are moved from the old URL, to the new URL. As
such, the GID merge process fails, because the URL has been prematurely deleted.

I've decided to manually do the GID merging in the URL DAO for now, by finding
old GIDs and updating merge tables, and then proceeding with the rest of the
merge.