MBS-3755: Ensure that the release editor artist field is always filled in from seed parameters
Seeding the release editor with an artist name (either as the artist name *or*
credited name) should always cause the artist name field to filled out. If only
the credited name was filled out for the release artist, the artist name would
remain blank (though this wasn't true for tracks which worked correctly).

This commit makes the field consistent no matter where it is used.