Fix URL controller tests
The url.sql test data was incorrect as it didn't specify entity types for the
end points of the test relationship. For data tests this was fine, as nothing
tried to load relationships, but the URL controller tests do load relationships.

I've made these columns NOT NULL and added a die statement to type_to_url if you
try to lookup the model for a type where there is no mapping.