Use a deferred constraint trigger to clean up unused URLs
By using a deferred constraint trigger we can allow the URL cleanup to only fire
at the end of a transaction. This is more correct, as by removing unused URLs we
are really trying to enforce the constraint that 'every URL row has one or more
relationships'. The current triggers enforced this, but they were too eager. For
example...

Take 2 URLs UA and UB, both with a relationship to the same label, L1. If we merge
UA into UB, we need to keep one of the relationships to L1. If we keep the L1-UB
link, then UA will get deleted as it has no relationships. We are now unable to
merge UB into UA, because UA doesn't exist! By deferring this constraint to the
end of the transaction, this problem goes away.