Selecting voting options should cause the background to change to reflect the selection
Adds a new bit of handling to the EditList javascript. When a user changes their
vote selection, the parent div's of each radio button are reset, and then have a
class applied depending on whether they are selected, and the type of vote they
represent. The edit stylesheet has been updated to have specific styling for
each of the "no", "yes" and "abstain" vote types.