Add some MusicBrainz specific Plack debug panels
This introduces a psgi script for running the server in a PSGI environment. The
script makes use of Plack and it's 'debug panel' support in order to provide
some debugging information right in the browser.

Of the default Plack debug panels, this enables the memory panel, for memory
information; the timer panel, for timing the request as a whole; and the request
variables panel for viewing query/body parameters.

I have also written 3 custom debug panels that should be useful when working on
getting the website fast. First, the DAOLogger logs time spent within data
access methods, showing the call tree along with timing information. Outliers
(past 1 standard deviation) are highlighted.

The Exclusive Time module is a more detailed look at the data methods, and also
includes calls to private methods. It shows the time spent exclusively within a
subroutine but *excluding* time spent calling other subroutines.

Finally, the Template Toolkit panel gives a brief overview of time spent within
actually doing the rendering, and time spent processing certain includes.