Display artists used in artist credits if the artists have comments
If a user is changing an artist credit with a single artist, and the artist
credit isn't changed (the credited name stays the same) the edit details would
not show the artist credit expanded. However, if the user is changing the actual
artist link, we still want to expand the artist credit so they can see the
artist comments.

This commit changes the requirements to show expanded artist credits to be:
* Credited name differs from artist name
* More than 1 credited artist
* Credited artist has a comment

This isn't ideal, as it now means that some artist credits might get expanded
when it's not strictly necessary. However, erring on the side of more rather
than less information seems to be useful here.