Fix the Selenium tests
The tests were failing because now that we have moved to DBIx::Connector, we
have thread/fork saftey. This means that if you fork, the child will get a new
database connection. However, we assumed that we didn't have this behavior. For
now, I've monkey patched DBIx::Connector to disable the thread saftey.