Be a little more intelligent on displaying the 'copy track changes' checkbox
The checkbox is now dynamically shown and hidden, when the user makes changes
via the recording association bubble. If they choose 'add a new recording' the
checkbox is hidden, and if they chose an existing recording, they checkbox will
always be shown. It would be best to check if there are actual changes, but that
requires serializing the recording into the form and seems more effort than it
is worth for now, and certainly more effort to just fix MBS-4185.

The initial form still does a check however, but the check has been extended to
hide the checkbox if the default option is to add a new recording - for example,
when adding a new release.