MBS-4211: Do not try and copy track changes to new recordings
When editing a release, users have the option of copying track changes to
recordings. However, this should only happen if the recording actually
exists. If the user is creating a new recording, then track changes should not
be copied. This currently causes an SQL exception to be thrown, and this commit
resolves that.