Use namespace::autoclean in data modules
This means that a subroutine is only available inside a certain package, and is
not exported out of the package. For example, we have a lot of private
functions, and they should only be callable from within that package. This is
especially important with the new instrumentation code I've added, which looks
at the list of exported methods and wraps them all. Some subroutines are
functions, not methods, so wrapping these is in correct (as attempts to pass
$self).

By using namespace::autoclean, these functions will not be incorrectly wrapped.