Only allow removing auto-editor rights if the checkbox is present and not ticked
There is some odd behavior if you use _insert_edit with a form that doesn't have
an "as auto editor" checkbox at all. The action assumes that the user intended
to *uncheck* this box, but they didn't have the choice and thus probably did
expect it to be entered as an autoedit.

In order to prevent this confusing and somewhat buggy behavior, this overriding
logic now only happens if the form has an 'as_auto_editor' checkbox.