Revert changes in bd4373e7ae35a
That commit changes the merging to fail if any of the releases current mediums
don't have a position set. However, this fails if a release has had mediums
added after the merge edit was entered.

The original behavior is to fail if any of the mediums that are being reordered
no longer exists.

There's certainly some sort of middle ground where it's safe to carry on
merging (without an error), but it's neither of these. Until I can figure out
exactly what that logic is, I'm reverting to have the edge case crash, rather
than the commen case.