MBS-4408: Clone in defaultSubmitAction retaining event handlers
The defaultSubmitAction stuff works by cloning the desired submit button and
prepending it to the form. However, a call to clone() does not copy event
handlers, by default. For the CAA work, we have a special bit of validation
hooked on submission which is integral to everything working properly, so we
need to copy this event handler.