MBS-3482: Cleanup a smaller subset of YouTube URLs
Now only cleans up YouTube urls that are in the /user/ path. The cleanup here is
to trim trailing path arguments (other than the user name), and to remove query
parameters and URL fragments. If the URL is not a /user/ URL, then the cleanup
does not happen and the URL is left unchanged. This is because nikki has spotted
a large collection of URLs which are incorrectly transformed, and it feels like
there are too many exceptions for this to ever make sense.