MBS-4566: Link to advanced query syntax
The 'advanced search' option should link to advanced query syntax, but the
changes in MBS-834 (commit dea316d) removed this link. To add the link, I've had
to inline the actual labels into the template, in order to be able to make use
of the doc_link macro, which is less that optimal but it does work.